Add Core seams for Tasks extension
- Add IMcpServerRawHandlerRegistry seam (McpRawRequestHandler delegate, registry over request handlers, IsDraftProtocolRequest) and McpServerOptions.RawRequestHandlerConfigurators
- Add McpServer.CurrentOutgoingRequestInterceptor AsyncLocal seam and McpOutgoingRequestInterceptor delegate; route SampleAsync/ElicitAsync/RequestRootsAsync through it
- Make McpSession.IsDraftProtocol public and McpClient.ResolveInputRequestsAsync public abstract (MCPEXP002)
- Remove Tasks coupling from Core: ConfigureTasks, task-store tools/call wrapper, SetTaskAugmented, CallToolWithTaskHandler/Filters, task lifecycle handlers, MaxConsecutiveStuckPolls, task method/notification/meta consts, task JsonSerializable entries
- Rewrite client CallToolAsync to a plain tools/call returning CallToolResult
- Relocate Tasks DTOs and store types into ModelContextProtocol.Extensions.Tasks; drop McpTaskExecutionContext
- Add ModelContextProtocol.Extensions.Tasks project and TasksJsonContext
