+++ src/ModelContextProtocol.Core/McpJsonUtilities.cs
@@ -120,12 +120,6 @@ public static partial class McpJsonUtilities
     [JsonSerializable(typeof(ResourceUpdatedNotificationParams))]
     [JsonSerializable(typeof(RootsListChangedNotificationParams))]
     [JsonSerializable(typeof(ToolListChangedNotificationParams))]
-    [JsonSerializable(typeof(TaskStatusNotificationParams))]
-    [JsonSerializable(typeof(WorkingTaskNotificationParams))]
-    [JsonSerializable(typeof(CompletedTaskNotificationParams))]
-    [JsonSerializable(typeof(FailedTaskNotificationParams))]
-    [JsonSerializable(typeof(CancelledTaskNotificationParams))]
-    [JsonSerializable(typeof(InputRequiredTaskNotificationParams))]
 
     // MCP Request Params / Results
     [JsonSerializable(typeof(CallToolRequestParams))]
@@ -173,19 +167,6 @@ public static partial class McpJsonUtilities
     [JsonSerializable(typeof(InputResponse))]
     [JsonSerializable(typeof(IDictionary<string, InputRequest>))]
     [JsonSerializable(typeof(IDictionary<string, InputResponse>))]
-
-    [JsonSerializable(typeof(GetTaskRequestParams))]
-    [JsonSerializable(typeof(GetTaskResult))]
-    [JsonSerializable(typeof(WorkingTaskResult))]
-    [JsonSerializable(typeof(CompletedTaskResult))]
-    [JsonSerializable(typeof(FailedTaskResult))]
-    [JsonSerializable(typeof(CancelledTaskResult))]
-    [JsonSerializable(typeof(InputRequiredTaskResult))]
-    [JsonSerializable(typeof(UpdateTaskRequestParams))]
-    [JsonSerializable(typeof(UpdateTaskResult))]
-    [JsonSerializable(typeof(CancelTaskRequestParams))]
-    [JsonSerializable(typeof(CancelTaskResult))]
-    [JsonSerializable(typeof(CreateTaskResult))]
 
     // MCP Content
     [JsonSerializable(typeof(ContentBlock))]
