+++ src/ModelContextProtocol.Core/Protocol/NotificationMethods.cs
@@ -145,16 +145,6 @@ public static class NotificationMethods
     public const string CancelledNotification = "notifications/cancelled";
 
     /// <summary>
-    /// The name of the notification sent by the server to push task status updates to subscribed clients.
-    /// </summary>
-    /// <remarks>
-    /// Part of the <c>io.modelcontextprotocol/tasks</c> extension.
-    /// Each notification carries a complete task state for the current status, identical to what
-    /// <c>tasks/get</c> would have returned at that moment.
-    /// </remarks>
-    public const string TaskStatusNotification = "notifications/tasks/status";
-
-    /// <summary>
     /// The name of the notification sent first on a <see cref="RequestMethods.SubscriptionsListen"/>
     /// response stream to indicate which notification types the server agreed to deliver.
     /// </summary>
