+++ src/ModelContextProtocol.Core/Protocol/Result.cs
@@ -30,8 +30,8 @@ public abstract class Result
     /// When absent or set to <c>"complete"</c>, the result is a normal completed response.
     /// When set to <c>"input_required"</c>, the result is an <see cref="InputRequiredResult"/> indicating
     /// that additional input is needed before the request can be completed.
-    /// When set to <c>"task"</c>, the result is a <see cref="CreateTaskResult"/> indicating that the server
-    /// has created a long-running task in lieu of returning the result directly.
+    /// When set to <c>"task"</c>, the server has created a long-running task in lieu of returning the
+    /// result directly.
     /// </para>
     /// </remarks>
     /// <value>Defaults to <see langword="null"/>, which is equivalent to <c>"complete"</c>.</value>
