+++ src/ModelContextProtocol.Core/Server/McpServerOptions.cs
@@ -191,17 +191,14 @@ public sealed class McpServerOptions
     public int MaxSamplingOutputTokens { get; set; } = 1000;
 
     /// <summary>
-    /// Gets or sets the task store for managing asynchronous task executions.
+    /// Gets or sets a list of configurators that register or wrap low-level raw request handlers
+    /// after the strongly-typed handlers have been configured.
     /// </summary>
     /// <remarks>
-    /// <para>
-    /// When set, the server automatically enables the <c>io.modelcontextprotocol/tasks</c> extension
-    /// and wires up <c>tasks/get</c>, <c>tasks/update</c>, and <c>tasks/cancel</c> handlers backed by this store.
-    /// Tool executions from clients that signal task support will be wrapped in tasks via the store.
-    /// </para>
-    /// <para>
-    /// If explicit task handlers are also set on <see cref="Handlers"/>, the explicit handlers take precedence.
-    /// </para>
+    /// This is an SDK extensibility hook intended for bolt-on packages (such as the MCP Tasks extension)
+    /// that need to register new request methods or wrap existing handlers. Each configurator receives an
+    /// <see cref="IMcpServerRawHandlerRegistry"/> for the server being constructed.
     /// </remarks>
-    public IMcpTaskStore? TaskStore { get; set; }
+    [Experimental(Experimentals.Subclassing_DiagnosticId, UrlFormat = Experimentals.Subclassing_Url)]
+    public IList<Action<IMcpServerRawHandlerRegistry>>? RawRequestHandlerConfigurators { get; set; }
 }
