+++ src/ModelContextProtocol.Extensions.Tasks/Protocol/CancelTaskResult.cs
