+++ src/ModelContextProtocol.Extensions.Tasks/Protocol/McpExtensions.cs
