+++ src/ModelContextProtocol.Extensions.Tasks/Protocol/McpTaskStatus.cs
