+++ src/ModelContextProtocol.Extensions.Tasks/Protocol/UpdateTaskResult.cs
