+++ src/ModelContextProtocol.Extensions.Tasks/Server/IMcpTaskStore.cs
