+++ src/ModelContextProtocol.Extensions.Tasks/Server/McpTaskInfo.cs
